fix: refactor edit initiative modal to correct filter persistence
