+++ src/components/Initiatives/InitiativeDetailsPage/InitativeFilterDialog/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { InitiativeFilterDialogWrapper as InitiativeFilterDialog } from './InitiativeFilterDialog';
+export { InitiativeFilterDialog as InitiativeFilterDialog } from './InitiativeFilterDialog';
